+++ REPOWildCardMod/Source/Patches/PlayerAvatarPatch.cs
@@ -5,7 +5,6 @@ namespace REPOWildCardMod.Patches
     [HarmonyPatch(typeof(PlayerAvatar))]
     public static class PlayerAvatarPatch
     {
-        static readonly BepInEx.Logging.ManualLogSource log = WildCardMod.log;
         [HarmonyPatch(nameof(PlayerAvatar.ChatMessageSend))]
         [HarmonyPostfix]
         public static void CheckForKill(PlayerAvatar __instance, ref string _message)